pkg/cli: move all table dump upload code to a separate file
The table dump upload code is expected to grow a lot more than other
artifact upload code. Moving this to it's own file will make it easier
to implement/review changes to this part of the code.

Epic: none
Release note: None
